Tests moved into separate files [ci skip]
